ADD Split de functionality of the library in server and client modules
