+++ test/unit/client-unregistration-test.js
@@ -23,7 +23,7 @@
 
 'use strict';
 
-var libLwm2m2 = require('../..'),
+var libLwm2m2 = require('../..').server,
     coap = require('coap'),
     utils = require('./testUtils'),
     config = require('../../config'),
