+++ test/unit/device-management-test.js
@@ -23,7 +23,7 @@
 
 'use strict';
 
-var libLwm2m2 = require('../..'),
+var libLwm2m2 = require('../..').server,
     utils = require('./testUtils'),
     config = require('../../config'),
     libcoap = require('coap'),
