+++ test/unit/testUtils.js
@@ -25,7 +25,7 @@
 
 /*jshint unused:false */
 
-var libLwm2m2 = require('../..'),
+var libLwm2m2 = require('../..').server,
     coap = require('coap'),
     Readable = require('stream').Readable,
     config = require('../../config'),
